Enabled input connections for simulation
Ports now can be :inputs or :registers, which determines the provenance
of the data.

Connections can be made to :registers or :subparts.

Connections are now just plain maps, not ASTNodes.

There isn't any "easy" api for getting subparts yet.

Most modules should have their *input-fns* and *input-fn-ports*
correctly constructed, which are necessary to have the logic and
registers needed to compute the inputs. This code isn't well tested and
probably doesn't work in all cases.